refactor(billing): give the payment-intent charge lifecycle a single owner
Move createPaymentIntent (the charge/replay state machine) and resolveTransaction
out of StripeService into a dedicated StripeTransactionService that owns the
payment-transaction records for user-initiated flows. The confirm-payment
controller and the wallet-reload job now call it, passing an explicit
onAmountMismatch policy ("reject" / "tolerate") instead of the service sniffing
the idempotency-key prefix. No change to charging or idempotency behavior.
